fix socket exception
